// packages/prettier-plugin-java/src/tree-sitter-java.d.ts
interface SyntaxNodeBase {
  tree: Tree;
  id: number;
  typeId: number;
  grammarId: number;
  type: string;
  grammarType: string;
  isNamed: boolean;
  isMissing: boolean;
  isExtra: boolean;
  hasChanges: boolean;
  hasError: boolean;
  isError: boolean;
  text: string;
  parseState: number;
  nextParseState: number;
  startPosition: Point;
  endPosition: Point;
  startIndex: number;
  endIndex: number;
  parent: SyntaxNode | null;
  children: Array<SyntaxNode>;
  namedChildren: Array<SyntaxNode>;
  childCount: number;
  namedChildCount: number;
  firstChild: SyntaxNode | null;
  firstNamedChild: SyntaxNode | null;
  lastChild: SyntaxNode | null;
  lastNamedChild: SyntaxNode | null;
  nextSibling: SyntaxNode | null;
  nextNamedSibling: SyntaxNode | null;
  previousSibling: SyntaxNode | null;
  previousNamedSibling: SyntaxNode | null;
  descendantCount: number;

  toString(): string;
  child(index: number): SyntaxNode | null;
  namedChild(index: number): SyntaxNode | null;
  childForFieldName(fieldName: string): SyntaxNode | null;
  childForFieldId(fieldId: number): SyntaxNode | null;
  fieldNameForChild(childIndex: number): string | null;
  childrenForFieldName(fieldName: string): Array<SyntaxNode>;
  childrenForFieldId(fieldId: number): Array<SyntaxNode>;
  firstChildForIndex(index: number): SyntaxNode | null;
  firstNamedChildForIndex(index: number): SyntaxNode | null;

  descendantForIndex(index: number): SyntaxNode;
  descendantForIndex(startIndex: number, endIndex: number): SyntaxNode;
  namedDescendantForIndex(index: number): SyntaxNode;
  namedDescendantForIndex(startIndex: number, endIndex: number): SyntaxNode;
  descendantForPosition(position: Point): SyntaxNode;
  descendantForPosition(startPosition: Point, endPosition: Point): SyntaxNode;
  namedDescendantForPosition(position: Point): SyntaxNode;
  namedDescendantForPosition(startPosition: Point, endPosition: Point): SyntaxNode;
  descendantsOfType<T extends TypeString>(types: T | readonly T[], startPosition?: Point, endPosition?: Point): NodeOfType<T>[];

  closest<T extends SyntaxType>(types: T | readonly T[]): NamedNode<T> | null;
  walk(): TreeCursor;
}
// ...
interface NamedNodeBase extends SyntaxNodeBase {
    isNamed: true;
}

/** An unnamed node with the given type string. */
export interface UnnamedNode<T extends string = string> extends SyntaxNodeBase {
  type: T;
  isNamed: false;
}

type PickNamedType<Node, T extends string> = Node extends { type: T; isNamed: true } ? Node : never;

type PickType<Node, T extends string> = Node extends { type: T } ? Node : never;

/** A named node with the given `type` string. */
export type NamedNode<T extends SyntaxType = SyntaxType> = PickNamedType<SyntaxNode, T>;
// ...
export type SimpleTypeNode =
  | BooleanTypeNode
  | FloatingPointTypeNode
  | GenericTypeNode
  | IntegralTypeNode
  | ScopedTypeIdentifierNode
  | TypeIdentifierNode
  | VoidTypeNode
  ;

export type TypeNode =
  | UnannotatedTypeNode
  | AnnotatedTypeNode
  ;

export type UnannotatedTypeNode =
  | SimpleTypeNode
  | ArrayTypeNode
  ;

export type DeclarationNode =
  | AnnotationTypeDeclarationNode
  | ClassDeclarationNode
  | EnumDeclarationNode
  | ImportDeclarationNode
  | InterfaceDeclarationNode
  | ModuleDeclarationNode
  | PackageDeclarationNode
  | RecordDeclarationNode
  ;

export type ExpressionNode =
  | AssignmentExpressionNode
  | BinaryExpressionNode
  | CastExpressionNode
  | InstanceofExpressionNode
  | LambdaExpressionNode
  | PrimaryExpressionNode
  | SwitchExpressionNode
  | TernaryExpressionNode
  | UnaryExpressionNode
  | UpdateExpressionNode
// ...
export interface ArrayInitializerNode extends NamedNodeBase {
  type: SyntaxType.ArrayInitializer;
}

export interface ArrayTypeNode extends NamedNodeBase {
  type: SyntaxType.ArrayType;
  dimensionsNode: DimensionsNode;
  elementNode: UnannotatedTypeNode;
}
// ...
export interface ClassBodyNode extends NamedNodeBase {
  type: SyntaxType.ClassBody;
}
// ...
export interface ConstantDeclarationNode extends NamedNodeBase {
  type: SyntaxType.ConstantDeclaration;
  declaratorNodes: VariableDeclaratorNode[];
  typeNode: UnannotatedTypeNode;
}
// ...
export interface DimensionsNode extends NamedNodeBase {
  type: SyntaxType.Dimensions;
}

export interface DimensionsExprNode extends NamedNodeBase {
  type: SyntaxType.DimensionsExpr;
}

export interface DoStatementNode extends NamedNodeBase {
  type: SyntaxType.DoStatement;
  bodyNode: StatementNode;
  conditionNode: ParenthesizedExpressionNode;
}

export interface ElementValueArrayInitializerNode extends NamedNodeBase {
  type: SyntaxType.ElementValueArrayInitializer;
}
// ...
export interface EnumBodyDeclarationsNode extends NamedNodeBase {
  type: SyntaxType.EnumBodyDeclarations;
}
// ...
export interface FieldDeclarationNode extends NamedNodeBase {
  type: SyntaxType.FieldDeclaration;
  declaratorNodes: VariableDeclaratorNode[];
  typeNode: UnannotatedTypeNode;
}
// ...
export interface InterfaceBodyNode extends NamedNodeBase {
  type: SyntaxType.InterfaceBody;
}
// ...
export interface LocalVariableDeclarationNode extends NamedNodeBase {
  type: SyntaxType.LocalVariableDeclaration;
  declaratorNodes: VariableDeclaratorNode[];
  typeNode: UnannotatedTypeNode;
}
// ...
export interface VariableDeclaratorNode extends NamedNodeBase {
  type: SyntaxType.VariableDeclarator;
  dimensionsNode?: DimensionsNode;
  nameNode: IdentifierNode | UnderscorePatternNode;
  valueNode?: ArrayInitializerNode | ExpressionNode;
}
// ...
export interface IdentifierNode extends NamedNodeBase {
  type: SyntaxType.Identifier;
}
// ...
export interface UnderscorePatternNode extends NamedNodeBase {
  type: SyntaxType.UnderscorePattern;
}